+++ backend/packages/sandbox/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @cat-factory/sandbox
 
+## 0.11.88
+
+### Patch Changes
+
+- Updated dependencies [1025674]
+- Updated dependencies [e5f7eb0]
+  - @cat-factory/contracts@0.263.0
+  - @cat-factory/agents@0.117.1
+  - @cat-factory/kernel@0.262.1
+  - @cat-factory/sandbox-fixtures@0.7.298
+
+## 0.11.87
+
+### Patch Changes
+
+- Updated dependencies [8cbd518]
+- Updated dependencies [8cbd518]
+- Updated dependencies [7a2730a]
+  - @cat-factory/contracts@0.262.0
+  - @cat-factory/kernel@0.262.0
+  - @cat-factory/agents@0.117.0
+  - @cat-factory/sandbox-fixtures@0.7.297
+
 ## 0.11.86
 
 ### Patch Changes